third_party/blink/web_tests/external/wpt/streams/count-queuing-strategy.any.worker-expected.txt

This is a testharness.js-based test.
PASS Can construct a CountQueuingStrategy with a valid high water mark
PASS Can construct a CountQueuingStrategy with any value as its high water mark
FAIL CountQueuingStrategy constructor behaves as expected with strange arguments assert_throws: construction fails with null function "() => new CountQueuingStrategy(null)" did not throw
FAIL CountQueuingStrategy.prototype.size should work generically on its this and its arguments Illegal invocation
FAIL CountQueuingStrategy size behaves as expected with strange arguments Illegal invocation
FAIL CountQueuingStrategy instances have the correct properties assert_object_equals: highWaterMark property should be a data property with the value passed the constructor value is undefined, expected object
FAIL CountQueuingStrategy's highWaterMark property can be set to anything Cannot assign to read only property 'highWaterMark' of object '#<CountQueuingStrategy>'
PASS CountQueuingStrategy.name is correct
PASS subclassing CountQueuingStrategy should work correctly
Harness: the test ran to completion.
